refactor: PreferenceRepository 패키지 위치를 repository로 변경
