+++ backend/src/main/java/com/jujeol/preference/domain/repository/PreferenceRepository.java
@@ -1,7 +1,9 @@
-package com.jujeol.preference.domain;
+package com.jujeol.preference.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
+
+import com.jujeol.preference.domain.Preference;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +11,6 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
 public interface PreferenceRepository extends JpaRepository<Preference, Long> {
-
 
     Optional<Preference> findByMemberIdAndDrinkId(Long memberId, Long drinkId);
