+++ backend/src/test/java/com/jujeol/member/domain/PreferenceRepositoryTest.java
@@ -13,7 +13,7 @@ import com.jujeol.member.auth.domain.ProviderName;
 import com.jujeol.member.member.domain.Member;
 import com.jujeol.member.member.domain.repository.MemberRepository;
 import com.jujeol.preference.domain.Preference;
-import com.jujeol.preference.domain.PreferenceRepository;
+import com.jujeol.preference.domain.repository.PreferenceRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
